apps/agents/requirements.txt
# Agno Multi-Agent Framework
agno==0.1.0

# LLM Providers
openai==1.61.0

# Web Framework
fastapi==0.115.8
uvicorn[standard]==0.34.0

# Database
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Environment
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.1

# Vector DB Support (for RAG)
pgvector==0.3.6

# Pydantic for type safety
pydantic==2.10.6

# Web search (for agents)
duckduckgo-search==7.3.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
respx==0.22.0
